\name{RclBLAS-package}
\alias{RclBLAS-package}
\alias{RclBLAS}
\docType{package}
\title{
\packageTitle{RclBLAS}
}
\description{
\packageDescription{RclBLAS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{RclBLAS}
\packageIndices{RclBLAS}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{RclBLAS}

Maintainer: \packageMaintainer{RclBLAS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
